fel: add support for Allwinner A733
The Allwinner A733 is built on a 12nm process and features a dual-core
ARM Cortex-A76 combined with a hexa-core ARM Cortex-A55 using the
big.LITTLE architecture.

The chip has 16K + 192K of SRAM-A2, with first 16K used only by CPUS
domain, later it follows by 512K of SHARED-SRAM with no gaps.

 SRAM_A2       0x40000 - 0x73FFF
 SHARED_SRAM   0x74000 - 0xF3FFF

The BootROM SRAM usage is similar to other SoCs: there is the IRQ stack
growing down from around 34K offset in SRAM-A2, and probably some
buffers located towards the end of the SRAM. In this setup, we've only
tried to use the 192K block of SRAM-A2, which should be enough for SPL
setup. It's possible to use more SRAM, for example, the SHARED-SRAM,
but we leave it for now.

This setup is known to boot the WIP mainline U-Boot setup, including
some placeholder TF-A port for now. SPL execution, including returning
back to the BROM, works, also the 64-bit switch, as well as the SID
dump.

Signed-off-by: Yixun Lan <dlan@gentoo.org>